:building_construction: Split TimeUtils to UTC and Instant class to handle MockBean within tests
