+++ backend/src/main/java/org/opensource/schulaltas/service/TimeUTC.java
@@ -3,7 +3,7 @@ package org.opensource.schulaltas.service;
 import org.springframework.stereotype.Service;
 
 @Service
-public class TimeUtil {
+public class TimeUTC {
 
  public Long now () {
   return System.currentTimeMillis();
